requirements.txt
Django==1.8
django-debug-toolbar==1.5
sqlparse==0.2.4
